# Conflicts:
#	src/components/Card/RecommendCard.js
#	src/components/Tour/AddTourDemo.js
#	src/utils/fetchData.js

+++ src/redux/callApi/postCall.js
@@ -16,7 +16,13 @@ export const getPosts = token => async dispatch => {
 
     // console.log(res.data.posts)
 
-    dispatch(postAction.getPosts({ posts: res.data.posts, id: 0 }));
+    dispatch(
+      postAction.getPosts({
+        posts: res.data.posts,
+        id: 0,
+        postId: res.data.postId
+      })
+    );
   } catch (err) {
     // console.log(err);
     dispatch(postAction.error({ error: 'Có lỗi xảy ra' }));
@@ -72,11 +78,14 @@ export const getMorePost = data => async dispatch => {
     const { postId } = data;
     const limit = postId.length;
 
-    const res = customAxios().post(`/post/list?detail=true&limit=${limit}`, {
-      postId: postId
-    });
+    const res = await customAxios().post(
+      `/post/list?detail=true&limit=${limit}`,
+      {
+        list: postId
+      }
+    );
 
-    dispatch(postAction.getMorePost({ posts: res }));
+    dispatch(postAction.getMorePost({ posts: res.data.posts }));
   } catch (err) {
     dispatch(postAction.error({ error: 'Có lỗi xảy ra' }));
   }
@@ -104,9 +113,7 @@ export const getPostById = (id, token, next) => async dispatch => {
   }
 };
 
-export const createPost =
-  (data, token, type, socket, next, error, createReview) => async dispatch => {
-    // console.log("go?")
+export const createPost = (data, token, type, socket, next, error, createReview) => async dispatch => {
     // post api
     try {
       let image = [];
@@ -207,17 +214,18 @@ export const likePost = (id, auth, socket, next) => async dispatch => {
     dispatch(postAction.updateLike({ id: id, likes: res.data.likes }));
     socket.emit('like', { type: 'post', id, likes: res.data.likes });
 
-    //notify
-    const dataNotify = {
-      id: auth.user._id,
-      text: ' thích bài viết của bạn',
-      recipients: [res.data.post.userId],
-      content: res.data.post.content,
-      image:
-        res.data.post.images.length > 0 ? res.data.post.images[0] : 'empty',
-      url: `/post/${id}`
-    };
-    dispatch(createNotify(dataNotify, auth.token, socket));
+    if(auth.user._id !== res.data.post.userId){
+        const dataNotify = {
+          id: auth.user._id,
+          text: ' thích bài viết của bạn',
+          recipients: [res.data.post.userId],
+          content: res.data.post.content,
+          image:
+            res.data.post.images.length > 0 ? res.data.post.images[0] : 'empty',
+          url: `/post/${id}`
+        };
+        dispatch(createNotify(dataNotify, auth.token, socket));
+    }
   } catch (err) {
     next();
     // console.log(err);
@@ -234,11 +242,13 @@ export const unlikePost = (id, auth, socket, next) => async dispatch => {
     socket.emit('unlike', { type: 'post', id, likes: res.data.likes });
 
     // Notify
-    const dataNotify = {
-      id: auth.user._id,
-      url: `/post/${id}`
-    };
-    dispatch(deleteNotify(dataNotify, auth.token, socket));
+    if(auth.user._id !== res.data.post.userId){
+        const dataNotify = {
+          id: auth.user._id,
+          url: `/post/${id}`
+        };
+        dispatch(deleteNotify(dataNotify, auth.token, socket));
+    }
   } catch (err) {
     next();
     if (err.response && err.response.data && err.response.data.message)
